refactor: 기능별 파일 분류
